Separated bob conversion helper functions into separate file
Makes component file clearer and ecapsulates it in one file.

Also makes clearer which utilities are actually used in the component vs being helper functions for the other utilities.

Also changed console logs to normal logs
